add yagit installation
