+++ cmake/Config.cmake.in
@@ -1,0 +1,11 @@
+@PACKAGE_INIT@
+
+include(CMakeFindDependencyMacro)
+find_dependency(GDCM REQUIRED)
+if(@GAMMA_VERSION@ STREQUAL "THREADED")
+    find_dependency(Threads REQUIRED)
+endif()
+
+include("${CMAKE_CURRENT_LIST_DIR}/@TARGETS_EXPORT_NAME@.cmake")
+
+check_required_components(@PROJECT_NAME@)
