+++ docs/source/authors.rst
@@ -1,0 +1,2 @@
+Authors
+=======
