+++ docs/source/index.rst
@@ -7,5 +7,6 @@ Yet Another Gamma Index Tool |release| documentation
    introduction
    installation
    algorithms
+   authors
    api/api
    examples
