+++ examples/gamma2DInterp.cpp
@@ -35,7 +35,7 @@
 #include <string>
 #include <iostream>
 
-#include <yagit.hpp>
+#include "yagit/yagit.hpp"
 
 int main(int argc, char** argv){
     if(argc <= 2){
