+++ include/yagit/GammaParameters.hpp
