+++ include/yagit/GammaResult.hpp
@@ -18,17 +18,30 @@
  ********************************************************************************************/
 #pragma once
 
-#include <string>
+#include "yagit/ImageData.hpp"
 
-#include "ImageData.hpp"
-
-namespace yagit::DataWriter{
+namespace yagit{
 
 /**
- * @brief Write image to MetaImage file (.mha)
- * @param img Image to save
- * @param filepath File path where image will be saved
+ * @brief Container storing gamma index values
  */
-void writeToMetaImage(const ImageData& img, const std::string& filepath);
+class GammaResult : public ImageData{
+public:
+    using ImageData::ImageData;
+
+    /// @brief The percentage of elements that are less than or equal to 1. It ignores NaN values.
+    value_type passingRate() const;
+
+    /// @brief Minimum value of gamma index. It ignores NaN values.
+    value_type minGamma() const;
+    /// @brief Maximum value of gamma index. It ignores NaN values.
+    value_type maxGamma() const;
+    /// @brief Sum of values of gamma index. It ignores NaN values.
+    double sumGamma() const;
+    /// @brief Mean of values of gamma index. It ignores NaN values.
+    double meanGamma() const;
+    /// @brief Variance of values of gamma index. It ignores NaN values.
+    double varGamma() const;
+};
 
 }
