+++ include/yagit/Image.hpp
